backend/requirements.txt
Pillow
Sanic
sanic-session
Levenshtein
opencv-python
asyncio
setuptools
sanic_cors